+++ src/components/Comment/Comment.scss
@@ -23,6 +23,7 @@
   top: 1px;
   border: 0;
   width: 50px;
+  height: 20px;
   font-size: 16px;
   background-color: transparent;
   color: black;
