+++ src/components/Feed/Feed.scss
@@ -1,0 +1,114 @@
+img {
+  margin: 0 4px;
+  cursor: pointer;
+}
+
+#feed-img {
+  margin: -10px 0;
+}
+
+.feeds {
+  * {
+    background-color: white;
+  }
+  width: 502px;
+  position: relative;
+  right: 100px;
+  margin: 30px auto;
+  border: 1px solid lightgray;
+  padding: 3px 0 0 0;
+  &-top {
+    display: flex;
+    justify-content: space-between;
+    padding: 17px 20px 23px;
+    margin: -3px 0;
+    &-left,
+    &-right {
+      display: flex;
+      align-items: center;
+    }
+    &-right {
+      cursor: pointer;
+    }
+  }
+}
+
+#profile-img {
+  border-radius: 30px;
+  width: 30px;
+  height: 30px;
+}
+
+.feeds-icons {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 20px 10px 10px;
+  &-left {
+    display: flex;
+    align-items: center;
+    div {
+      padding: 0 5px;
+    }
+  }
+}
+
+.feeds-likes {
+  display: flex;
+  align-items: center;
+  padding: 0 0 0 10px;
+}
+
+.feeds-comments-top {
+  position: relative;
+  left: 20px;
+  padding-bottom: 5px;
+}
+
+.fa-heart {
+  cursor: pointer;
+}
+
+.time {
+  color: gray;
+  position: relative;
+  padding-bottom: 5px;
+}
+
+.feed-time {
+  padding-left: 20px;
+}
+
+.nickname {
+  font-weight: 800;
+  cursor: pointer;
+}
+
+.write-comment {
+  display: flex;
+  background-color: white;
+  border-top: 1px solid lightgray;
+  input {
+    width: 70%;
+    outline: none;
+    border: none;
+    background-color: white;
+    padding: 0 0 0 10px;
+  }
+  button {
+    width: 100px;
+    font-size: 16px;
+    background-color: white;
+    border: none;
+    color: lightblue;
+    margin: 12px 0 0 200px;
+  }
+}
+
+@media screen and (max-width: 900px) {
+  .feeds {
+    position: relative;
+    right: 0;
+    margin: 20px auto;
+  }
+}
