+++ src/pages/jinwoo/Main/Main.scss
@@ -1,5 +1,10 @@
 * {
-  background-color: white;
+  background-color: #f9f9f9;
+}
+
+.main {
+  position: relative;
+  margin: 70px 0 0 0;
 }
 
 .navigation {
@@ -8,10 +13,9 @@
   justify-content: space-around;
   align-items: center;
   font-size: 30px;
-  height: 55px;
+  height: 70px;
   background-color: white;
   padding: 30px;
-  margin: 80px 0 0 0;
 }
 
 .nav-left,
@@ -23,11 +27,12 @@
 
 .nav-center {
   position: relative;
-  background-color: transparent;
 }
 
 #search-icon {
   background-color: transparent;
+  position: relative;
+  bottom: 5px;
 }
 
 .fa-magnifying-glass {
@@ -38,11 +43,18 @@
 }
 
 #search {
-  padding: 0 40px;
+  padding: 0 20px;
+  height: 40px;
 }
 
-.nav-right img {
-  cursor: pointer;
+.nav-right {
+  img {
+    cursor: pointer;
+    background-color: white;
+  }
+  div {
+    background-color: white;
+  }
 }
 
 #user {
@@ -173,10 +185,6 @@ img {
   cursor: pointer;
 }
 
-#feed-img {
-  margin: 0;
-}
-
 #division {
   background: black;
   width: 2px;
@@ -192,70 +200,10 @@ main {
   padding: 30px;
 }
 
-.feeds {
-  width: 500px;
-  position: relative;
-  right: 100px;
-  margin: 0 auto;
-  border: 1px solid lightgray;
-  padding: 3px 0;
-}
-
-.feeds-top {
-  display: flex;
-  justify-content: space-between;
-  margin: 20px;
-}
-
-.feeds-top-left,
-.feeds-top-right {
-  display: flex;
-  align-items: center;
-}
-
-.feeds-top-right {
-  cursor: pointer;
-}
-
-#profile-img {
-  border-radius: 30px;
-  width: 30px;
-  height: 30px;
-}
-
-.feeds-icons {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin: 10px;
-}
-
-.feeds-icons-left {
-  display: flex;
-  align-items: center;
-}
-
-.feeds-icons-left > div {
-  padding: 0 5px;
-}
-
-.feeds-likes {
-  display: flex;
-  align-items: center;
-  padding: 0 0 0 10px;
-}
-
-.feeds-comments-top {
-  position: relative;
-  left: 20px;
-  padding-bottom: 5px;
-}
-
 .fa-heart {
   cursor: pointer;
 }
 
-.time,
 .others,
 .ad-desc,
 .story-text,
@@ -263,54 +211,27 @@ main {
   color: gray;
 }
 
+.time {
+  background-color: white;
+}
+
 .others {
   padding-bottom: 5px;
   width: 140px;
-}
-
-.time {
-  position: relative;
-  padding-bottom: 5px;
-}
-
-.feed-time {
-  padding-left: 20px;
+  background-color: white;
 }
 
 .nickname {
   font-weight: 800;
   cursor: pointer;
-}
-
-.write-comment {
-  display: flex;
   background-color: white;
-  border-top: 1px solid lightgray;
-  margin: 3px 0;
-  z-index: -1;
-  input {
-    width: 70%;
-    outline: none;
-    border: none;
-    background-color: white;
-    padding: 0 0 0 10px;
-  }
-  button {
-    width: 100px;
-    font-size: 16px;
-    background-color: white;
-    border: none;
-    color: lightblue;
-    margin: 12px 0 0 200px;
-  }
 }
 
 .main-right {
-  position: fixed;
+  position: absolute;
   top: 0;
-  left: 66vw;
+  left: 70vw;
   margin: 150px auto;
-  background-color: rgb(249, 249, 249);
   width: 400px;
 }
 
@@ -341,6 +262,10 @@ main {
   border: 1px solid lightgray;
   margin: 30px;
   padding: 10px;
+  background-color: white;
+  &-text {
+    background-color: white;
+  }
 }
 
 .story-top {
@@ -348,20 +273,27 @@ main {
   width: 300px;
   padding: 10px;
   justify-content: space-between;
+  background-color: white;
 }
 
 .story-user {
   display: flex;
   padding: 5px;
+  background-color: white;
+  div {
+    background-color: white;
+  }
 }
 
 .all {
   cursor: pointer;
+  background-color: white;
 }
 
 .suggestion {
   min-width: 320px;
   border: 1px solid lightgray;
+  background-color: white;
   margin: 30px;
   padding: 10px;
 }
@@ -370,11 +302,16 @@ main {
   display: flex;
   padding: 10px;
   justify-content: space-between;
+  background-color: white;
 }
 
 .suggestion-user {
   display: flex;
   padding: 5px;
+  background-color: white;
+  div {
+    background-color: white;
+  }
 }
 
 .follow-btn {
@@ -400,10 +337,5 @@ main {
 @media screen and (max-width: 900px) {
   .main-right {
     display: none;
-  }
-  .feeds {
-    position: relative;
-    right: 0;
-    margin: 0 auto;
   }
 }
